system/MsgDialog: RAII for MsgDialog ui
